requirement.txt
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.3
django-import-export==2.7.1
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
tablib==3.2.0
xlrd==2.0.1
xlwt==1.3.0
django-cors-headers==3.11.0
pyjwt
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
